Refactored folder structure, added Scroll to Top button
