+++ frontend/src/features/permits/context/ApplicationContext.tsx
@@ -1,4 +1,4 @@
-import { Dispatch, createContext } from "react";
+import { Dispatch, ReactElement, createContext } from "react";
 import { TermOversizeApplication } from "../types/application";
 
 interface ApplicationContextType {
@@ -7,6 +7,11 @@ interface ApplicationContextType {
   next: () => void;
   back: () => void;
   goTo: (index: number) => void;
+  steps?: ReactElement;
+  currentStepIndex?: number;
+  step?: ReactElement;
+  isFirstStep?: boolean;
+  isLastStep?: boolean;
 }
 
 export const ApplicationContext = createContext<ApplicationContextType>({
