+++ frontend/src/features/permits/hooks/hooks.ts
@@ -2,43 +2,12 @@ import { useQueryClient, useMutation } from "@tanstack/react-query";
 import { useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import { SnackBarContext } from "../../../App";
-import { submitTermOversize } from "./permitsAPI";
+import { submitTermOversize } from "../apiManager/permitsAPI";
 
-import { ReactElement, useState } from "react";
-
-export const useMultiStepForm = (steps: ReactElement[]) => {
-  const [currentStepIndex, setCurrentStepIndex] = useState(0);
-
-  const next = () => {
-    setCurrentStepIndex((i) => {
-      if (i >= steps.length - 1) return i;
-      return i + 1;
-    });
-  };
-
-  const back = () => {
-    setCurrentStepIndex((i) => {
-      if (i <= 0) return i;
-      return i - 1;
-    });
-  };
-
-  const goTo = (index: number) => {
-    setCurrentStepIndex(index);
-  };
-
-  return {
-    currentStepIndex,
-    step: steps[currentStepIndex],
-    steps,
-    isFirstStep: currentStepIndex === 0,
-    isLastStep: currentStepIndex === steps.length - 1,
-    goTo,
-    next,
-    back,
-  };
-};
-
+/**
+ * A custom react query mutation hook that submits the application data to the backend API
+ * then shows a snackbar and navigates back a page
+ */
 export const useSubmitTermOversizeMutation = () => {
   const queryClient = useQueryClient();
   const snackBar = useContext(SnackBarContext);
