+++ frontend/src/features/permits/hooks/useMultiStepForm.tsx
@@ -1,0 +1,39 @@
+import { ReactElement, useState } from "react";
+
+/**
+ * A custom hook to manage navigating the different steps of the application process by using react useState
+ * @param steps an array of React Elements, which are the steps of the application process
+ * @returns a list of methods and variables to navigate between steps of the application process
+ */
+export const useMultiStepForm = (steps: ReactElement[]) => {
+  const [currentStepIndex, setCurrentStepIndex] = useState(0);
+
+  const next = () => {
+    setCurrentStepIndex((i) => {
+      if (i >= steps.length - 1) return i;
+      return i + 1;
+    });
+  };
+
+  const back = () => {
+    setCurrentStepIndex((i) => {
+      if (i <= 0) return i;
+      return i - 1;
+    });
+  };
+
+  const goTo = (index: number) => {
+    setCurrentStepIndex(index);
+  };
+
+  return {
+    currentStepIndex,
+    step: steps[currentStepIndex],
+    steps,
+    isFirstStep: currentStepIndex === 0,
+    isLastStep: currentStepIndex === steps.length - 1,
+    goTo,
+    next,
+    back,
+  };
+};
