+++ frontend/src/features/permits/pages/TermOversize/TermOversize.scss
@@ -1,7 +1,6 @@
-@import "../../../../../themes/orbcStyles";
+@import "../../../../themes/orbcStyles.scss";
 
 //TODO: not yet implemented
-
 .tros-main-box {
   display: flex;
   flex-wrap: wrap;
