+++ frontend/src/features/permits/pages/TermOversize/formComponents/ConditionsTable.tsx
