+++ frontend/src/features/permits/pages/TermOversize/formComponents/PermitDetails.tsx
